feat(ui): remove redundant header chrome, delegate background to host
Ninth slice of the Figma redesign stack, continuing the topbar work
in #79 (stacked on it).

- Remove the "Show summary" toggle and the pass-rate/passed/failed/
  not-run summary panel entirely (LLMTestRunnerSummary deleted).
  computeSummaryStats moves into test-case-filtering.ts, where it's
  still needed for the toolbar's filter-pill counts.
- Remove "Import suite" and "Export suite" from the header — both are
  now redundant with the toolbar's own icon buttons. Remove "Add
  question" from the header too, for the same reason (the toolbar's
  "+" already covers it).
- Remove the header's divider (border-bottom/box-shadow) and the
  outer container's border/border-radius, so the header, toolbar, and
  test-case list read as one continuous surface instead of stacked
  panels.
- <llm-test-runner> no longer sets any background color anywhere
  (:host, .test-runner-container, the header, the toolbar, the list)
  — that's now entirely up to whatever page it's embedded in, matching
  how the Figma design applies its tray color at the page level, not
  per-component.
- Demo: give #runner-host the Figma page-level background
  (var(--tailwind-colors-slate-200, #e2e8f0)) as an example of a host
  page opting into that look, full-bleed to the actual viewport edges
  (left/right/bottom) rather than stopping at the page's own padded
  content width. Along the way, fixed two content-box/box-sizing bugs
  this surfaced (.container overflowing body's own right padding, and
  body's min-height: 100vh not including its own top+bottom padding,
  both from adding width/height-affecting properties without
  box-sizing: border-box) — added a universal box-sizing reset to
  prevent the same class of bug recurring.
