+++ src/components/llm-test-runner/header/llm-test-runner-header.css
@@ -1,21 +1,17 @@
 /* Test Runner Header Block */
 .test-runner-header {
   display: flex;
-  justify-content: space-between;
+  justify-content: flex-end;
   align-items: center;
   gap: var(--spacing-4);
   padding: var(--spacing-3) var(--spacing-5);
-  background: var(--background);
   color: var(--foreground);
-  border-bottom: var(--border-width) solid var(--border);
   position: sticky;
   top: var(--llmtr-sticky-top, 0px);
   z-index: var(--z-sticky);
-  box-shadow: var(--shadow-sm);
 }
 
 /* Header Elements */
-.test-runner-header__left,
 .test-runner-header__right {
   display: flex;
   gap: var(--spacing-3);
@@ -43,43 +39,6 @@
   line-height: var(--line-height-normal);
 }
 
-/* Modifier: Hidden */
-.test-runner-header--hidden {
-  display: none;
-}
-
-/* "Show summary" checkbox toggle */
-.test-runner-header__toggle {
-  display: inline-flex;
-  align-items: center;
-  gap: var(--spacing-2);
-  cursor: pointer;
-  user-select: none;
-  padding: var(--spacing-1) var(--spacing-2);
-  border-radius: var(--radius-md);
-  color: var(--muted-foreground);
-  transition: color 150ms ease, background-color 150ms ease;
-}
-
-.test-runner-header__toggle:hover {
-  color: var(--foreground);
-  background: var(--muted);
-}
-
-.test-runner-header__toggle-input {
-  width: 16px;
-  height: 16px;
-  margin: 0;
-  cursor: pointer;
-  accent-color: var(--primary);
-}
-
-.test-runner-header__toggle-label {
-  font-size: var(--font-size-sm);
-  font-weight: var(--font-weight-medium);
-  letter-spacing: var(--letter-spacing-snug);
-}
-
 /* Responsive Design */
 @media (max-width: 768px) /* var(--breakpoint-md) */ {
   .test-runner-header {
@@ -88,7 +47,6 @@
     padding: var(--spacing-4);
   }
 
-  .test-runner-header__left,
   .test-runner-header__right {
     flex-wrap: wrap;
     justify-content: center;
