--- src/components/llm-test-runner/summary/llm-test-runner-summary.css
@@ -1,148 +1,0 @@
-/* ------------------------------------------------------------------ */
-/* Summary bar — sits directly beneath the top header as an extension. */
-/* ------------------------------------------------------------------ */
-
-.test-runner-summary {
-  display: flex;
-  align-items: center;
-  gap: var(--spacing-6);
-  padding: var(--spacing-3) var(--spacing-5);
-  background: var(--background);
-  color: var(--foreground);
-  border-bottom: var(--border-width) solid var(--border);
-  position: sticky;
-  /* Sits just below the main header (which is also sticky at top:0). The
-   * header is ~57px tall on default sizing; using a calc keeps us tied to
-   * the header padding/line-height tokens so it stays accurate. */
-  top: var(--summary-bar-offset, 57px);
-  z-index: calc(var(--z-sticky) - 1);
-}
-
-/* Pass-rate cluster — large numeric reading on the far left. */
-.test-runner-summary__rate {
-  display: flex;
-  align-items: baseline;
-  gap: var(--spacing-2);
-}
-
-.test-runner-summary__rate-value {
-  font-size: var(--font-size-2xl);
-  font-weight: var(--font-weight-bold);
-  letter-spacing: var(--letter-spacing-tight);
-  color: var(--foreground);
-  line-height: var(--line-height-none);
-}
-
-.test-runner-summary__rate-label {
-  font-size: var(--font-size-sm);
-  color: var(--muted-foreground);
-}
-
-/* Status chips */
-.test-runner-summary__chips {
-  display: flex;
-  align-items: center;
-  gap: var(--spacing-3);
-}
-
-.test-runner-summary__chip {
-  display: inline-flex;
-  align-items: center;
-  gap: 6px;
-  padding: 4px 12px;
-  border-radius: var(--radius-full);
-  font-size: var(--font-size-sm);
-  font-weight: var(--font-weight-medium);
-  border: var(--border-width) solid transparent;
-  white-space: nowrap;
-}
-
-.test-runner-summary__chip::before {
-  content: '';
-  width: 7px;
-  height: 7px;
-  border-radius: 50%;
-  background: currentColor;
-  flex-shrink: 0;
-}
-
-.test-runner-summary__chip--passed {
-  background: var(--success-soft-bg);
-  color: var(--success-soft-fg);
-  border-color: var(--success-soft-border);
-}
-
-.test-runner-summary__chip--failed {
-  background: var(--destructive-soft-bg);
-  color: var(--destructive-soft-fg);
-  border-color: var(--destructive-soft-border);
-}
-
-.test-runner-summary__chip--not-run {
-  background: var(--muted);
-  color: var(--muted-foreground);
-  border-color: var(--border);
-}
-
-/* Trailing total — quiet, sits to the right of the chips. */
-.test-runner-summary__total {
-  font-size: var(--font-size-sm);
-  color: var(--muted-foreground);
-  margin-left: auto;
-}
-
-/* ------------------------------------------------------------------ */
-/* Number transitions                                                  */
-/* Each numeric span is re-keyed when its value changes, which causes  */
-/* Stencil to swap the node and re-run the entrance animation. Subtle  */
-/* pop + fade so changes feel "live" without being distracting.        */
-/* ------------------------------------------------------------------ */
-.test-runner-summary__rate-value,
-.test-runner-summary__chip-count,
-.test-runner-summary__total-count {
-  display: inline-block;
-  transform-origin: center;
-  animation: test-runner-summary-number-pop 320ms
-    cubic-bezier(0.34, 1.56, 0.64, 1);
-}
-
-.test-runner-summary__rate-value {
-  /* Slightly slower + larger overshoot for the headline number. */
-  animation-duration: 420ms;
-}
-
-@keyframes test-runner-summary-number-pop {
-  0% {
-    transform: scale(0.7) translateY(3px);
-    opacity: 0;
-  }
-  55% {
-    transform: scale(1.12) translateY(0);
-    opacity: 1;
-  }
-  100% {
-    transform: scale(1) translateY(0);
-    opacity: 1;
-  }
-}
-
-@media (prefers-reduced-motion: reduce) {
-  .test-runner-summary__rate-value,
-  .test-runner-summary__chip-count,
-  .test-runner-summary__total-count {
-    animation: none;
-  }
-}
-
-/* Responsive */
-@media (max-width: 768px) /* var(--breakpoint-md) */ {
-  .test-runner-summary {
-    flex-wrap: wrap;
-    gap: var(--spacing-3);
-    padding: var(--spacing-3) var(--spacing-4);
-  }
-
-  .test-runner-summary__total {
-    margin-left: 0;
-  }
-}
