+++ src/components/llm-test-runner/test-cases/test-case-filtering.ts
@@ -3,10 +3,50 @@ import { computeTestStatus } from './llm-test-case-row';
 
 export type StatusFilter = 'all' | 'not-tested' | 'failed' | 'passed';
 
+export interface SummaryStats {
+  total: number;
+  passed: number;
+  failed: number;
+  notRun: number;
+  passRate: number;
+}
+
 /**
- * Buckets match the ones computeSummaryStats already uses: 'not-tested'
- * covers both 'not-run' and the transient 'running' state, 'failed'
- * covers both 'failed' and 'partial'.
+ * Aggregate per-test statuses into summary counts.
+ *
+ * Buckets:
+ *  - passed  → fully passed
+ *  - failed  → has results but didn't fully pass (includes 'partial')
+ *  - notRun  → no completed evaluation yet (includes currently running)
+ *
+ * passRate is passed / total, rounded to nearest integer.
+ */
+export function computeSummaryStats(testCases: TestCase[]): SummaryStats {
+  const total = testCases.length;
+  let passed = 0;
+  let failed = 0;
+  let notRun = 0;
+
+  for (const tc of testCases) {
+    const { kind } = computeTestStatus(tc);
+    if (kind === 'passed') {
+      passed += 1;
+    } else if (kind === 'failed' || kind === 'partial') {
+      failed += 1;
+    } else {
+      // 'not-run' | 'running' | 'evaluating'
+      notRun += 1;
+    }
+  }
+
+  const passRate = total > 0 ? Math.round((passed / total) * 100) : 0;
+  return { total, passed, failed, notRun, passRate };
+}
+
+/**
+ * Buckets match computeSummaryStats: 'not-tested' covers both 'not-run'
+ * and the transient 'running' state, 'failed' covers both 'failed' and
+ * 'partial'.
  */
 export function filterTestCasesByStatus(
   testCases: TestCase[],
