+++ src/components/llm-test-runner/test-cases/test-cases-toolbar.css
@@ -3,16 +3,16 @@
 /* quick status filters, and search, sitting above the test-case list. */
 /* ------------------------------------------------------------------ */
 
-/* Shares .test-cases's background + horizontal padding so the two sit
- * on one continuous tray with no seam, and the toolbar's content lines
- * up with the cards below it. */
+/* Shares .test-cases's horizontal padding so the toolbar's content lines
+ * up with the cards below it. No background of its own — this sits on
+ * whatever background the host page provides, same as the rest of the
+ * component (see llm-test-runner.css/:host). */
 .test-cases-toolbar {
   display: flex;
   align-items: center;
   justify-content: space-between;
   gap: var(--spacing-2);
   padding: var(--spacing-3) var(--spacing-5) var(--spacing-2);
-  background: var(--muted);
 }
 
 .test-cases-toolbar__left {
@@ -52,8 +52,9 @@
 }
 
 /* Inactive pills use --background (white), not --secondary — --secondary
- * resolves to the same value as this toolbar's --muted background in
- * this token set, which would make an inactive pill invisible. */
+ * resolves to the same value as --muted in this token set, which is a
+ * likely host-page background for this toolbar and would make an
+ * inactive pill invisible against it. */
 .test-cases-toolbar__filter-pill {
   display: inline-flex;
   align-items: center;
